HSEARCH-3222 Better usage of DocumentReferenceExtractorHelper
