+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/search/extraction/impl/DocumentReferenceExtractorHelper.java
@@ -6,6 +6,8 @@
  */
 package org.hibernate.search.backend.lucene.search.extraction.impl;
 
+import java.util.Set;
+
 import org.apache.lucene.document.Document;
 import org.hibernate.search.backend.lucene.search.impl.LuceneDocumentReference;
 import org.hibernate.search.backend.lucene.util.impl.LuceneFields;
@@ -14,6 +16,15 @@ import org.hibernate.search.engine.search.DocumentReference;
 public final class DocumentReferenceExtractorHelper {
 
 	private DocumentReferenceExtractorHelper() {
+	}
+
+	public static void contributeCollectors(LuceneCollectorsBuilder luceneCollectorBuilder) {
+		luceneCollectorBuilder.requireTopDocsCollector();
+	}
+
+	public static void contributeFields(Set<String> absoluteFieldPaths) {
+		absoluteFieldPaths.add( LuceneFields.indexFieldName() );
+		absoluteFieldPaths.add( LuceneFields.idFieldName() );
 	}
 
 	public static DocumentReference extractDocumentReference(Document document) {
